Implement col and row in moves history
