+++ src/components/MovesHistory/styles.scss
@@ -1,0 +1,20 @@
+.table {
+  display: table;
+}
+
+.row {
+  display: table-row;
+}
+
+.col {
+  display: table-cell;
+}
+
+.move {
+  background-color: #ffffff90;
+  padding: 10px;
+  border: none;
+  border-radius: 5px;
+  font-family: "Goldman", cursive;
+  margin: 0.2rem;
+}
