+++ src/pages/Game/styles.scss
@@ -3,15 +3,6 @@ ul {
   padding-left: 30px;
 }
 
-.move {
-  background-color: #ffffff90;
-  padding: 10px;
-  border: none;
-  border-radius: 15px;
-  font-family: "Goldman", cursive;
-  margin: 0.3rem;
-}
-
 .status {
   margin-bottom: 10px;
 }
